+++ cli/src/commands/feature-flag/commands/delete.ts
@@ -1,8 +1,6 @@
 import { Command } from 'commander';
 import pc from 'picocolors';
-import { EnumStatusCode } from '@wundergraph/cosmo-connect/dist/common/common_pb';
 import inquirer from 'inquirer';
-import Table from 'cli-table3';
 import ora from 'ora';
 import { BaseCommandOptions } from '../../../core/types/types.js';
 import { getBaseHeaders } from '../../../core/config.js';
@@ -15,6 +13,11 @@ export default (opts: BaseCommandOptions) => {
   command.option('-n, --namespace [string]', 'The namespace of the feature flag.');
   command.option('-f --force', 'Flag to force the deletion (skip confirmation).');
   command.option('--suppress-warnings', 'This flag suppresses any warnings produced by composition.');
+  command.option(
+    '--disable-resolvability-validation',
+    'This flag will disable the validation for whether all nodes of the federated graph are resolvable. Do NOT use unless troubleshooting.',
+  );
+
   command.action(async (name, options) => {
     if (!options.force) {
       const deletionConfirmed = await inquirer.prompt({
@@ -32,6 +35,7 @@ export default (opts: BaseCommandOptions) => {
 
     const resp = await opts.client.platform.deleteFeatureFlag(
       {
+        disableResolvabilityValidation: options.disableResolvabilityValidation,
         name,
         namespace: options.namespace,
       },