+++ protographic/tests/sdl-to-proto/02-complex-types.test.ts
@@ -36,13 +36,18 @@ describe('SDL to Proto - Complex Types', () => {
         rpc QueryUsersByRole(QueryUsersByRoleRequest) returns (QueryUsersByRoleResponse) {}
       }
 
+      // Wrapper message for a list of User.
+      message ListOfUser {
+        repeated User items = 1;
+      }
+
       // Request message for usersByRole operation.
       message QueryUsersByRoleRequest {
         UserRole role = 1;
       }
       // Response message for usersByRole operation.
       message QueryUsersByRoleResponse {
-        repeated User users_by_role = 1;
+        ListOfUser users_by_role = 1;
       }
 
       message User {
@@ -93,6 +98,8 @@ describe('SDL to Proto - Complex Types', () => {
       "syntax = "proto3";
       package service.v1;
 
+      import "google/protobuf/wrappers.proto";
+
       // Service definition for DefaultService
       service DefaultService {
         rpc MutationCreateUser(MutationCreateUserRequest) returns (MutationCreateUserResponse) {}
@@ -104,7 +111,7 @@ describe('SDL to Proto - Complex Types', () => {
       }
       // Response message for dummy operation.
       message QueryDummyResponse {
-        string dummy = 1;
+        google.protobuf.StringValue dummy = 1;
       }
       // Request message for createUser operation.
       message MutationCreateUserRequest {
@@ -118,14 +125,14 @@ describe('SDL to Proto - Complex Types', () => {
       message UserInput {
         string name = 1;
         string email = 2;
-        int32 age = 3;
+        google.protobuf.Int32Value age = 3;
       }
 
       message User {
         string id = 1;
         string name = 2;
         string email = 3;
-        int32 age = 4;
+        google.protobuf.Int32Value age = 4;
       }"
     `);
   });
@@ -222,6 +229,11 @@ describe('SDL to Proto - Complex Types', () => {
         rpc QueryRootNode(QueryRootNodeRequest) returns (QueryRootNodeResponse) {}
       }
 
+      // Wrapper message for a list of TreeNode.
+      message ListOfTreeNode {
+        repeated TreeNode items = 1;
+      }
+
       // Request message for rootNode operation.
       message QueryRootNodeRequest {
       }
@@ -242,7 +254,7 @@ describe('SDL to Proto - Complex Types', () => {
         string id = 1;
         string value = 2;
         TreeNode parent = 3;
-        repeated TreeNode children = 4;
+        ListOfTreeNode children = 4;
       }"
     `);
   });
@@ -284,9 +296,21 @@ describe('SDL to Proto - Complex Types', () => {
       "syntax = "proto3";
       package service.v1;
 
+      import "google/protobuf/wrappers.proto";
+
       // Service definition for DefaultService
       service DefaultService {
         rpc QueryUsers(QueryUsersRequest) returns (QueryUsersResponse) {}
+      }
+
+      // Wrapper message for a list of AddressInput.
+      message ListOfAddressInput {
+        repeated AddressInput items = 1;
+      }
+
+      // Wrapper message for a list of User.
+      message ListOfUser {
+        repeated User items = 1;
       }
 
       // Request message for users operation.
@@ -295,14 +319,14 @@ describe('SDL to Proto - Complex Types', () => {
       }
       // Response message for users operation.
       message QueryUsersResponse {
-        repeated User users = 1;
+        ListOfUser users = 1;
       }
 
       message UserFilterInput {
-        string name_contains = 1;
-        int32 min_age = 2;
-        int32 max_age = 3;
-        repeated AddressInput addresses = 4;
+        google.protobuf.StringValue name_contains = 1;
+        google.protobuf.Int32Value min_age = 2;
+        google.protobuf.Int32Value max_age = 3;
+        ListOfAddressInput addresses = 4;
       }
 
       message User {
@@ -315,7 +339,7 @@ describe('SDL to Proto - Complex Types', () => {
         string street = 1;
         string city = 2;
         string country = 3;
-        string zip_code = 4;
+        google.protobuf.StringValue zip_code = 4;
       }"
     `);
   });