+++ studio/CHANGELOG.md
@@ -4,6 +4,22 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.122.2](https://github.com/wundergraph/cosmo/compare/studio@0.122.1...studio@0.122.2) (2025-07-21)
+
+**Note:** Version bump only for package studio
+
+## [0.122.1](https://github.com/wundergraph/cosmo/compare/studio@0.122.0...studio@0.122.1) (2025-07-18)
+
+### Bug Fixes
+
+* **studio:** fix broken doc links ([#2052](https://github.com/wundergraph/cosmo/issues/2052)) ([2fd6f40](https://github.com/wundergraph/cosmo/commit/2fd6f4017eee0cba6344cc99e4dc768689931bda)) (@Blacksmoke16)
+
+# [0.122.0](https://github.com/wundergraph/cosmo/compare/studio@0.121.0...studio@0.122.0) (2025-07-10)
+
+### Features
+
+* add the `subgraph-viewer` role ([#2022](https://github.com/wundergraph/cosmo/issues/2022)) ([902b3bc](https://github.com/wundergraph/cosmo/commit/902b3bcd8ae9efaab71f5e1a97012938af51a57f)) (@wilsonrivera)
+
 # [0.121.0](https://github.com/wundergraph/cosmo/compare/studio@0.120.2...studio@0.121.0) (2025-07-09)
 
 ### Features